requirements.txt
flask
flask_mysqldb
flask_sqlalchemy
flask_restful
flask-bcrypt
werkzeug.*
flask_mail
itsdangerous
jwt
flask_jwt_extended == 4.3.1
PyJWT == 2.3.0